refactor(model-config): move workspace model state into store
